Added functionality to get temperature data for a specific city using OpenWeatherMap API

+++ RideService.Services/Entities/Main.cs
@@ -6,6 +6,6 @@ namespace RideService.Services.Entities
 {
     public class Main
     {
-        public decimal Temp { get; set; }
+        public double Temp { get; set; }
     }
 }
